Handle shield and potions as methods on Player
This change gets us easier consitency around when a shield should break
or when a player can validly drink a potion. It means that the game
state logic can simply hand off cards to the player and not worry about
the state logic governing if a potion will heal or fizzle, etc.